♻️ (#344) 댓글 액션 로직 및 유틸 분리
