+++ src/features/task-detail/components/CommentSection/CommentSection.tsx
@@ -1,15 +1,7 @@
-import toast from 'react-hot-toast';
-import type { CommentUIType } from '@/features/comment/types/commentTypes';
-import { useCreateCommentMutation } from '@/features/comment/hooks/useCreateCommentMutation';
-import { useUpdateCommentMutation } from '@/features/comment/hooks/useUpdateCommentMutation';
-import { useDeleteCommentMutation } from '@/features/comment/hooks/useDeleteCommentMutation';
-import { useTaskDetailQuery } from '@/features/task/hooks/useTaskDetailQuery';
-import { useTaskDetailStore } from '@/features/task-detail/store/taskDetail/useTaskDetailStore';
-import { useCommentSelect } from '@/features/task-detail/hooks/useCommentSelect';
-import type { FileInfo } from '@/features/task-detail/types/taskDetailType';
 import CommentList from '@/features/task-detail/components/CommentSection/CommentList';
 import CommentEditor from '@/features/task-detail/components/CommentSection/CommentEditor';
-import { useShallow } from 'zustand/react/shallow';
+import type { CommentUIType } from '@/features/comment/types/commentTypes';
+import { useCommentActions } from '@/features/task-detail/hooks/useCommentActions';
 
 interface CommentSectionProps {
   projectId: string;
@@ -18,64 +10,10 @@ interface CommentSectionProps {
 }
 
 const CommentSection = ({ projectId, taskId, comments }: CommentSectionProps) => {
-  const { mutate: createComment } = useCreateCommentMutation(projectId, taskId);
-  const { mutate: updateComment } = useUpdateCommentMutation(projectId, taskId);
-  const { mutate: deleteComment } = useDeleteCommentMutation(projectId, taskId);
-
-  const { data: task } = useTaskDetailQuery(projectId, taskId);
-  const { pins, clearCurrentPin, currentPin, persona } = useTaskDetailStore(
-    useShallow((s) => ({
-      pins: s.pins,
-      clearCurrentPin: s.clearCurrentPin,
-      currentPin: s.currentPin,
-      persona: s.persona,
-    })),
+  const { handlePinClick, handleCreate, handleUpdate, handleDelete } = useCommentActions(
+    projectId,
+    taskId,
   );
-  const { commentSelect } = useCommentSelect();
-  /** 댓글의 핀 아이콘 클릭 시 PDF 위치 이동 */
-  const handlePinClick = (fileInfo: FileInfo | null) => {
-    if (!fileInfo || !task?.files) return;
-    commentSelect(fileInfo, task.files, pins);
-  };
-
-  /** 댓글 등록 */
-  const handleCreate = (data: {
-    content: string;
-    isAnonymous: boolean;
-    fileInfo?: FileInfo | null;
-  }) => {
-    if (!data.content.trim()) return toast.error('댓글을 입력해주세요!');
-    const resolvedFileInfo = data.fileInfo ?? currentPin ?? null;
-
-    const newComment = {
-      content: data.content,
-      isAnonymous: data.isAnonymous,
-      persona: persona,
-      ...(resolvedFileInfo ? { fileInfo: resolvedFileInfo } : {}),
-    };
-    createComment({ commentData: newComment });
-    clearCurrentPin();
-  };
-
-  /** 댓글 수정 */
-  const handleUpdate = (
-    commentId: string,
-    data: { content: string; isAnonymous: boolean; fileInfo?: FileInfo | null },
-  ) => {
-    if (!data.content.trim()) return toast.error('댓글을 입력해주세요!');
-
-    const { content, isAnonymous, fileInfo } = data;
-    const updatedData = {
-      content,
-      isAnonymous,
-      ...(fileInfo !== undefined ? { fileInfo } : {}),
-    };
-
-    updateComment({ commentId, updatedData });
-  };
-
-  /** 댓글 삭제 */
-  const handleDelete = (commentId: string) => deleteComment(commentId);
 
   return (
     <div className="bg-gray-100 border-none flex flex-col h-full relative">
